Enum osztályok paramétereinek változtatása, képek finomhangolása.

+++ src/model/board/element/character/GhostMonster.java
@@ -3,7 +3,7 @@ import model.board.Board;
 
 import java.awt.*;
 public class GhostMonster extends Monster {
-    public GhostMonster(int x, int y, int width, int height, int velocity, Image image, boolean alive, boolean visible, Board board) {
+    public GhostMonster(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Board board) {
         super(x, y, width, height, velocity, image, alive, visible, board);
     }
