+++ src/model/board/element/character/IntelligentMonster.java
@@ -5,7 +5,7 @@ import model.board.Board;
 import java.awt.*;
 
 public class IntelligentMonster extends Monster {
-    public IntelligentMonster(int x, int y, int width, int height, int velocity, Image image, boolean alive, boolean visible, Board board) {
+    public IntelligentMonster(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Board board) {
         super(x, y, width, height, velocity, image, alive, visible, board);
     }
