+++ src/model/board/element/character/Monster.java
@@ -8,7 +8,7 @@ import java.awt.*;
 public abstract class Monster extends Entity {
     protected Board board;
 
-    public Monster(int x, int y, int width, int height, int velocity, Image image, boolean alive, boolean visible, Board board) {
+    public Monster(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Board board) {
         super(x, y, width, height, velocity, image, alive, visible);
         this.board = board;
     }
