+++ src/model/board/element/character/SemiIntelligentMonster.java
@@ -6,7 +6,7 @@ import java.awt.*;
 
 public class SemiIntelligentMonster extends Monster {
 
-    public SemiIntelligentMonster(int x, int y, int width, int height, int velocity, Image image, boolean alive, boolean visible, Board board) {
+    public SemiIntelligentMonster(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Board board) {
         super(x, y, width, height, velocity, image, alive, visible, board);
     }
     public Player getClosestPlayer(){ return null;}
