+++ src/model/board/element/deposable/Box.java
@@ -6,7 +6,7 @@ import java.awt.*;
 public class Box extends Entity {
     private Bonus bonus;
 
-    public Box(int x, int y, int width, int height, int velocity, Image image, boolean alive, boolean visible, Bonus bonus) {
+    public Box(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Bonus bonus) {
         super(x, y, width, height, velocity, image, alive, visible);
         this.bonus = bonus;
     }
