+++ src/model/board/element/powerup/benefit/GhostBonus.java
@@ -6,7 +6,7 @@ import model.board.element.powerup.BonusWithTimer;
 import java.awt.*;
 
 public class GhostBonus extends BonusWithTimer {
-    public GhostBonus(int x, int y, int width, int height, int velocity, Image image, boolean alive, boolean visible, Player owner) {
+    public GhostBonus(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible, Player owner) {
         super(x, y, width, height, velocity, image, alive, visible, owner);
     }
